inGameui component
